Moved transform to OrientationViewer, it should probably be deleted soon
